Initial commit
This is almost surely just going to be a rewrite of
__jamesshore/quixote__ but it intrigues me nonetheless. There was enough
about that package's API that I didn't like that I decided to do a port.
It's also the first chance that I have to try out the seed project. This
commit is just some stuff thrown together, nothing cohesive thus far.